feat(stm32mp1): allow to configure the serial boot load address
For product with 128MB DDR size, the OP-TEE is located at the end of the
DDR and the FIP can't be loaded at the default location because it overlap
the OP-TEE final location. So the default value for DWL_BUFFER_BASE is
invalid.

To avoid this conflict the serial boot load address = DWL_BUFFER_BASE
can be modified with a configuration flags.

Signed-off-by: Patrick Delaunay <patrick.delaunay@foss.st.com>
Change-Id: Ie27b87c10c57fea5d4c6200ce4f624e775b9a080
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/241601
Reviewed-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Yann GAUTIER <yann.gautier@foss.st.com>