fix(st-drivers): exit fconf_populate_mce if chip doesn't support crypto
The chip versions 135A and 135D do not support crypto and secure boot.
As fconf_populate_mce is called whatever the version is used, we should
exit the function without error for these chips. Else there is an
error because the st,mem-encrypt node doesn't exist, and the chip won't
be able to setup DDR encryption with MCE.

Signed-off-by: Yann Gautier <yann.gautier@foss.st.com>
Change-Id: I7c1b7e39998077b1e7e22a37511db48deae71002
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/244177
Reviewed-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Nicolas LE BAYON <nicolas.le.bayon@st.com>
Reviewed-by: CIBUILD <MDG-smet-aci-builds@list.st.com>