refactor(fdts): remove unused secure-status properties
For peripheral where both status and secure-status are set to okay,
the function fdt_get_status() returns the same status (DT_SHARED) if
secure-status property is omitted. This secure-status property (set to
disabled) can then be removed in the SoC file, and also in board file
where it is set to okay. This is the case for GPIOZ, HASH, I2C4&6, IWDG2,
RNG and RTC. The same is done for other IPs not used in current boards
like CRYPT, USART1 and SPI6.

Signed-off-by: Yann Gautier <yann.gautier@foss.st.com>
Change-Id: I9f9360842d4d41288db0cf1b92063f347c72d137
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/246308
Reviewed-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>