fix(st): fix the retry mechanism for MTD devices
The current check was made on a limited block number which
is not enough flexible to manage all the NAND specifications.
Prefer to manage a maximum size of a partition to find the backup
partition. Change nand_bkp_offset to nand_block_sz for better
understanding.

Signed-off-by: Lionel Debieve <lionel.debieve@foss.st.com>
Change-Id: I75e4dea24393f3829f4b319f8ddc9787ebe2b904
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/266481
Reviewed-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Yann GAUTIER <yann.gautier@foss.st.com>
Reviewed-by: Christophe KERELLO <christophe.kerello@st.com>
Domain-Review: Yann GAUTIER <yann.gautier@foss.st.com>